@@ -1,27 +1,37 @@
 {
-    "name": "ibrows/hmac-bundle",
-    "type": "symfony-bundle",
-    "description": "Symfony HMAC Bundle adds a SecurityListenerFactor which provide a hmac authentication for configured firewall",
-    "keywords": ["HMAC","Security","SecurityListenerFactor","Symonfy2"],
-    "homepage": "http://www.ibrows.ch",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Marc Steiner",
-            "homepage": "https://github.com/ibrows"
-        }
-    ],
-    "minimum-stability": "stable",
-    "require": {
-        "symfony/security-bundle": "~2.7|~3.0|~4.4",
-        "symfony/framework-bundle": "~2.7|~3.0|~4.4",
-        "acquia/http-hmac-php": "~1.0"
-    },
-    "require-dev": {
-      "phpunit/phpunit": "~4.0",
-      "symfony/browser-kit": "~2.7|~4.4"
-    },
-    "autoload": {
-        "psr-4": { "Ibrows\\HmacBundle\\": "" }
+  "name": "ibrows/hmac-bundle",
+  "type": "symfony-bundle",
+  "description": "Symfony HMAC Bundle adds a SecurityListenerFactor which provide a hmac authentication for configured firewall",
+  "keywords": [
+    "HMAC",
+    "Security",
+    "SecurityListenerFactor",
+    "Symonfy5"
+  ],
+  "homepage": "http://www.ibrows.ch",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Marc Steiner",
+      "homepage": "https://github.com/ibrows"
     }
+  ],
+  "minimum-stability": "stable",
+  "require": {
+    "php": ">=8.0",
+    "symfony/security-bundle": "~5.4",
+    "symfony/framework-bundle": "~5.4",
+    "acquia/http-hmac-php": "^1.0",
+    "symfony/config": "^6.3",
+    "symfony/yaml": "^6.3"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "~10.0",
+    "symfony/browser-kit": "~5.4"
+  },
+  "autoload": {
+    "psr-4": {
+      "Ibrows\\HmacBundle\\": ""
+    }
+  }
 }
