@@ -2,71 +2,37 @@
 
 namespace Ibrows\HmacBundle\DependencyInjection\Security\Factory;
 
-use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
+use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\AuthenticatorFactoryInterface;
 use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\DefinitionDecorator;
 use Symfony\Component\DependencyInjection\Reference;
 
-class HmacFactory implements SecurityFactoryInterface
+
+class HmacFactory implements AuthenticatorFactoryInterface
 {
 
-    /**
-     * @param ContainerBuilder $container
-     * @param string           $id
-     * @param array            $config
-     * @param string           $userProvider serviceId
-     * @param  string          $defaultEntryPoint
-     * @return array
-     */
-    public function create(ContainerBuilder $container, $id, $config, $userProvider, $defaultEntryPoint)
-    {
-        $providerBaseId = 'ibrows_hmac.security.authentication.provider';
-        $providerId = $providerBaseId . '.' . $id;
-        $service = $container->setDefinition($providerId, new ChildDefinition($providerBaseId));
-        $service->replaceArgument(0, new Reference($userProvider));
-        $service->replaceArgument(2, $id);
-        $service->replaceArgument(4, $config['authentication_provider_key']);
-
-        $listenerBaseId = 'ibrows_hmac.security.authentication.listener';
-        $listenerId = $listenerBaseId . '.' . $id;
-        $service = $container->setDefinition($listenerId, new ChildDefinition($listenerBaseId));
-        $service->replaceArgument(0, new Reference('security.token_storage')); // need to avoid ServiceCircularReferenceException
-        $service->replaceArgument(2, $id);
-        $service->replaceArgument(3, $config['authentication_provider_key']);
-        $service->replaceArgument(4, $defaultEntryPoint);
-
-        return array($providerId, $listenerId, $defaultEntryPoint);
-    }
-
-
-    /**
-     * @return string
-     */
-    public function getKey()
+    public function getKey(): string
     {
         return 'ibrows_hmac';
     }
 
-
-    /**
-     * @return string
-     */
-    public function getPosition()
+    public function addConfiguration(NodeDefinition $builder): void
     {
-        return 'pre_auth';
-    }
-
-    /**
-     * @param NodeDefinition $node
-     */
-    public function addConfiguration(NodeDefinition $node)
-    {
-        $node
+        $builder
             ->children()
             ->scalarNode('authentication_provider_key')->defaultValue('ibrows')->end()
             ->end();
     }
 
+    public function createAuthenticator(ContainerBuilder $container, string $firewallName, array $config, string $userProviderId): string
+    {
+        $providerBaseId = 'ibrows_hmac.security.authenticator';
+        $providerId = $providerBaseId . '.' . $firewallName;
+        $service = $container->setDefinition($providerId, new ChildDefinition($providerBaseId));
+        $service->replaceArgument(0, new Reference($userProviderId));
+        $service->replaceArgument(1, $config['authentication_provider_key']);
+
+        return $providerId;
+    }
 }
